[refactor] extract all dummy datas to api file for future replacement, extract all types except shadcn component to type folder.
