+++ app/(Main)/dashboard/helpers.tsx
@@ -9,38 +9,55 @@ import {
   GoalProgressData,
 } from "@/types/chart"
 import dynamic from "next/dynamic"
+import { ReactElement } from "react"
 
 // Dynamically import components
 const ProfitChart = dynamic(() => import("@/components/dashboard/ProfitChart"))
-const RealizedPnlChart = dynamic(() => import("@/components/dashboard/RealizedPnlChart"))
-const TradeFundBase = dynamic(() => import("@/components/dashboard/TradeFundBase"))
+const RealizedPnlChart = dynamic(
+  () => import("@/components/dashboard/RealizedPnlChart"),
+)
+const TradeFundBase = dynamic(
+  () => import("@/components/dashboard/TradeFundBase"),
+)
 const TradeLog = dynamic(() => import("@/components/dashboard/TradeLog"))
-const TradeSummary = dynamic(() => import("@/components/dashboard/TradeSummary"))
+const TradeSummary = dynamic(
+  () => import("@/components/dashboard/TradeSummary"),
+)
 const TradePlan = dynamic(() => import("@/components/dashboard/TradePlan"))
-const GoalProgress = dynamic(() => import("@/components/dashboard/GoalProgress"), { ssr: false })
+const GoalProgress = dynamic(
+  () => import("@/components/dashboard/GoalProgress"),
+  { ssr: false },
+)
 
+// Define chart components map
+type ChartComponentProps = {
+  ProfitChart: ProfitChartData[];
+  RealizedPnlChart: PnLChartData[];
+  TradeFundBase: TradeFundData[];
+  TradeLog: Log[];
+  TradeSummary: TradeSummaryData;
+  TradePlan: Plan[];
+  GoalProgress: GoalProgressData[];
+};
+
+const chartComponentsMap = new Map<keyof ChartComponentProps, (props: any) => ReactElement>([
+  ["ProfitChart", (data) => <ProfitChart data={data} />],
+  ["RealizedPnlChart", (pnLData) => <RealizedPnlChart pnLData={pnLData} />],
+  ["TradeFundBase", (fundBase) => <TradeFundBase fundBase={fundBase} />],
+  ["TradeLog", (logs) => <TradeLog logs={logs} />],
+  ["TradeSummary", (summary) => <TradeSummary summary={summary} />],
+  ["TradePlan", (plans) => <TradePlan plans={plans} />],
+  ["GoalProgress", (progress) => <GoalProgress progress={progress} />],
+]);
 export const renderChart = (
   chartId: keyof DynamicChartProps,
   chartProps: DynamicChartProps,
-) => {
-  const props = chartProps[chartId]
+): ReactElement => {
+  const renderComponent = chartComponentsMap.get(chartId)
 
-  switch (chartId) {
-    case "ProfitChart":
-      return <ProfitChart data={props as ProfitChartData[]} />
-    case "RealizedPnlChart":
-      return <RealizedPnlChart pnLData={props as PnLChartData[]} />
-    case "TradeFundBase":
-      return <TradeFundBase fundBase={props as TradeFundData[]} />
-    case "TradeLog":
-      return <TradeLog logs={props as Log[]} />
-    case "TradeSummary":
-      return <TradeSummary summary={props as TradeSummaryData} />
-    case "TradePlan":
-      return <TradePlan plans={props as Plan[]} />
-    case "GoalProgress":
-      return <GoalProgress progress={props as GoalProgressData[]} />
-    default:
-      return <div>No chart found</div>
+  if (!renderComponent) {
+    return <div>No chart found</div>
   }
+
+  return renderComponent(chartProps[chartId])
 }
