+++ app/(Main)/dashboard/page.tsx
@@ -15,8 +15,9 @@ import { Delete,  History,  Pin, PinOff } from "lucide-react"
 import ButtonEffect from "@/components/buttons/ButtonEffect"
 import { motion } from "motion/react"
 import { renderChart } from "./helpers"
-import { DynamicChartProps, ChartID, ChartProps, CompactType, DashboardItem, ResponsiveLayouts } from "@/types/chart"
+import { CompactType, DashboardItem, ResponsiveLayouts } from "@/types/chart"
 import Toolbox from "@/components/dashboard/Toolbox"
+import { userLayouts, chartProps } from "@/api"
 
 const ResponsiveReactGridLayout = WidthProvider(Responsive)
 
@@ -28,187 +29,6 @@ const defaultProps = {
 }
 // * Layout => A single layoutItem config.
 // * Layouts => { [breakpoints]: Layout[] }
-
-// TODO:Replace dummy data
-const userLayouts: {
-  layouts: ResponsiveLayouts
-  toolbox: ResponsiveLayouts
-  compactType: CompactType
-} = {
-  layouts: DEFAULT_LAYOUTS,
-  toolbox: DEFAULT_TOOLBOX,
-  compactType: DEFAULT_COMPACTTYPE,
-}
-const chartProps: DynamicChartProps = {
-  ProfitChart: [
-    { month: "Jan", TWSE: 23, Me: 8, 相對表現: ((8 / 23) * 100).toFixed(1) },
-    { month: "Feb", TWSE: 27, Me: 10, 相對表現: ((10 / 27) * 100).toFixed(1) },
-    { month: "Mar", TWSE: 32, Me: 18, 相對表現: ((18 / 32) * 100).toFixed(1) },
-    { month: "Apr", TWSE: 35, Me: 20, 相對表現: ((20 / 35) * 100).toFixed(1) },
-    { month: "May", TWSE: 30, Me: 22, 相對表現: ((22 / 30) * 100).toFixed(1) },
-    { month: "Jun", TWSE: 28, Me: 24, 相對表現: ((24 / 28) * 100).toFixed(1) },
-  ],
-  RealizedPnlChart: [
-    { week: "June/1", 成交筆數: 118, 報酬率: 3, 獲利筆數: 78 },
-    { week: "June/2", 成交筆數: 11, 報酬率: 16.2, 獲利筆數: 4 },
-    { week: "June/3", 成交筆數: 10, 報酬率: -2, 獲利筆數: 1 },
-    { week: "June/4", 成交筆數: 20, 報酬率: 0.21, 獲利筆數: 10 },
-  ],
-  TradeFundBase: [
-    { name: "現金水位", value: 200000 },
-    { name: "持倉部位", value: 800000 },
-  ],
-  TradeLog: [
-    {
-      _id: "1",
-      type: "多單",
-      action: "現股買進",
-      target: { symbol: "AAPL", name: "Apple. Inc" },
-      date: "12:09",
-      price: 150.25,
-      quantity: 1000,
-      comment: "在開盤時買入",
-    },
-    {
-      _id: "2",
-      type: "多單",
-      action: "現股賣出",
-      target: { symbol: "AAPL", name: "Apple. Inc" },
-      date: "09:10",
-      price: 155.75,
-      quantity: 100,
-      comment: "在高點賣出",
-    },
-    {
-      _id: "3",
-      type: "多單",
-      action: "融資買進",
-      target: { symbol: "GOOGL", name: "Google. corp" },
-      date: "11:20",
-      price: 2700.5,
-      quantity: 50,
-      comment: "使用保證金進行買進",
-    },
-    {
-      _id: "4",
-      type: "多單",
-      action: "融資賣出",
-      target: { symbol: "GOOGL", name: "Google. corp" },
-      date: "12:32",
-      price: 2725.0,
-      quantity: 50,
-      comment: "持有一天後賣出",
-    },
-    {
-      _id: "5",
-      type: "空單",
-      action: "沖買",
-      target: { symbol: "TSLA", name: "Tsela. Inc" },
-      date: "09:55",
-      price: 900.0,
-      quantity: 20,
-    },
-    {
-      _id: "6",
-      type: "空單",
-      action: "沖賣",
-      target: { symbol: "TSLA", name: "Tsela. Inc" },
-      date: "09:50",
-      price: 910.0,
-      quantity: 20,
-    },
-  ],
-  TradeSummary: {
-    asset: {
-      _id: "1",
-      totalCost: 1000000,
-      totalMarketPrice: 3000000,
-      position: [
-        { asset_id: "2486", asset_name: "一詮", quantity: 4000, cost: 137 },
-        { asset_id: "8103", asset_name: "瀚荃", quantity: 4000, cost: 71 },
-        { asset_id: "9958", asset_name: "世紀剛", quantity: 1000, cost: 223 },
-        { asset_id: "3037", asset_name: "欣興", quantity: 1000, cost: 160 },
-      ],
-    },
-    currentPrices: [
-      { symbol: "2486", name: "一詮", closePrice: 128 },
-      { symbol: "8103", name: "瀚荃", closePrice: 55.1 },
-      { symbol: "9958", name: "世紀剛", closePrice: 208 },
-      { symbol: "3037", name: "欣興", closePrice: 161.5 },
-    ],
-  },
-  TradePlan: [
-    {
-      _id: "123456789",
-      type: "多單",
-      target: {
-        symbol: "AAPL",
-        name: "Apple Inc.",
-      },
-      action: "建倉",
-      entryPrice: 113,
-      targetPrice: 150,
-      stop: {
-        type: "停損",
-        price: 100,
-      },
-      expectation: 150 / 100,
-      isExecuted: false,
-    },
-    {
-      _id: "987654321",
-      type: "空單",
-      target: {
-        symbol: "TSLA",
-        name: "Tesla, Inc.",
-      },
-      action: "建倉",
-      entryPrice: 1000,
-      targetPrice: 1194,
-      stop: {
-        type: "停損",
-        price: 900,
-      },
-      expectation: 1194 / 900,
-      isExecuted: false,
-      comment: "some comment",
-    },
-    {
-      _id: "123456780",
-      type: "多單",
-      target: {
-        symbol: "AAPL",
-        name: "Apple Inc.",
-      },
-      action: "出場",
-      entryPrice: 155,
-      targetPrice: 180,
-      stop: { type: "停利", price: 150 },
-      expectation: 180 / 150,
-      isExecuted: true,
-    },
-  ],
-  GoalProgress : [
-    {
-      name: "Goal",
-      value: 100,
-      color: "var(--chart-3)",
-      actualValue: 1000000,
-    },
-    {
-      name: "已達成進度",
-      value: 40,
-      color: "var(--chart-2",
-      actualValue: 400000,
-    },
-    {
-      name: "剩餘進度",
-      value: 60,
-      color: "var(--chart-1",
-      actualValue: 600000,
-    },
-  ],
-}
 
 const initLayoutsState = (
   initialLayouts: ResponsiveLayouts,
