+++ components/BlurFade.tsx
@@ -5,26 +5,10 @@ import {
   AnimatePresence,
   motion,
   useInView,
-  UseInViewOptions,
   Variants,
 } from "motion/react"
+import { BlurFadeProps } from "@/types/shared"
 
-type MarginType = UseInViewOptions["margin"]
-
-interface BlurFadeProps {
-  children: React.ReactNode
-  className?: string
-  variant?: {
-    hidden: { y: number }
-    visible: { y: number }
-  }
-  duration?: number
-  delay?: number
-  yOffset?: number
-  inView?: boolean
-  inViewMargin?: MarginType
-  blur?: string
-}
 
 export default function BlurFade({
   children,
