+++ components/blog/ArticleSidebar.tsx
@@ -6,32 +6,8 @@ import Filter from "../Filter"
 import { PlaceholdersAndVanishInput } from "../form/PlaceholdersAndVanishInput"
 import AvatarGallery from "./AvatarGallery"
 import ButtonLink from "../buttons/ButtonLink"
+import { persons } from "@/api"
 
-// !If replace this dummy data,  remove remotePatterns in nextConfig
-const persons = [
-  {
-    id: "126542",
-    url: "https://images.unsplash.com/photo-1543610892-0b1f7e6d8ac1?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8NHx8YXZhdGFyfGVufDB8fDB8fHww",
-    name: "Adrian Paul",
-    description: "COO & Co-Founder",
-    tags: ["Floral", "Highlands", "Wildflowers", "Colorful", "Resilience"],
-  },
-
-  {
-    id: "765752",
-    url: "https://images.unsplash.com/photo-1494790108377-be9c29b29330?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8M3x8YXZhdGFyfGVufDB8fDB8fHww",
-    name: "Flualy Cual",
-    description: "Founder & CEO",
-    tags: ["Twilight", "Peaks", "Silhouette", "Evening Sky", "Peaceful"],
-  },
-  {
-    id: "45673",
-    url: "https://images.unsplash.com/photo-1438761681033-6461ffad8d80?w=500&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8Nnx8YXZhdGFyfGVufDB8fDB8fHww",
-    name: "Naymur Rahman",
-    description: "CTO & Co-Founder",
-    tags: ["Rocky", "Ridges", "Contrast", "Adventure", "Clouds"],
-  },
-]
 
 export default function ArticleSidebar() {
   return (
