+++ components/form/TypingAnimation.tsx
@@ -3,12 +3,7 @@
 import { useEffect, useState } from "react"
 
 import { cn } from "@/lib/utils"
-
-interface TypingAnimationProps {
-  text: string
-  duration?: number
-  className?: string
-}
+import { TypingAnimationProps } from "@/types/shared"
 
 export default function TypingAnimation({
   text,
